Refactoring: Split monolithic agents.py into modular files and add Copernicus Marine data tool
