+++ src/tools/reflection_tools.py
@@ -1,0 +1,67 @@
+# src/tools/reflection_tools.py
+import base64
+import os
+import logging
+from pydantic import BaseModel, Field
+from langchain_core.tools import StructuredTool
+from openai import OpenAI
+from ..config import API_KEY
+
+# Define the function to encode the image
+def encode_image(image_path):
+    with open(image_path, "rb") as image_file:
+        return base64.b64encode(image_file.read()).decode('utf-8')
+
+def reflect_on_image(image_path: str) -> str:
+    if not os.path.exists(image_path):
+        return f"Error: The file {image_path} does not exist."
+
+    base64_image = encode_image(image_path)
+
+    prompt = """You are a professional reviewer of scientific images. Your task is to provide feedback to the visual creator agent so they can improve their visualization. Provide constructive criticism and specific improvement suggestions where necessary. Evaluate the provided image using the following criteria:
+
+1. Axis and Font Quality: Evaluate the visibility of axes and appropriateness of font size and style. Are the axes clearly visible and labeled? Is the font legible and suitable for the image size?
+2. Label Clarity: Assess if labels are well-positioned and not overlapping. Are all labels clearly readable and properly placed?
+3. Color Scheme: Analyze the color choices. Is the color scheme appropriate for the data presented? Are the colors distinguishable and not causing visual confusion?
+4. Data Representation: Evaluate how well the data is represented. Are data points clearly visible? Is the chosen chart or graph type appropriate for the data?
+5. Legend and Scale: Check the presence and clarity of legends and scales. Are they present when necessary and easy to understand?
+6. Overall Layout: Assess the overall layout and use of space. Is the image well-balanced and visually appealing?
+7. Technical Issues: Identify any technical problems such as pixelation, blurriness, or artifacts that might affect the image quality.
+8. Scientific Accuracy: To the best of your ability, comment on whether the image appears scientifically accurate and free from obvious errors or misrepresentations.
+9. **Convention Adherence**: Verify that the figure follows scientific conventions. For example, when depicting variables like 'Depth of water' or other vertical dimensions, these should appear on the Y-axis with minimum values at the top and maximum depth at the bottom. This is a critically important scientific convention - if depth/vertical dimensions are incorrectly presented on the horizontal X-axis, assign a significantly lower score (1/10) and provide clear instructions for correction.
+
+Please provide a structured review addressing each of these points. Conclude with an overall assessment of the image quality, highlighting any significant issues or exemplary aspects. Finally, give the image a score out of 10, where 10 is perfect quality and 1 is unusable.
+"""
+    openai_client = OpenAI(api_key=API_KEY)
+    response = openai_client.chat.completions.create(
+        model="gpt-4o",
+        messages=[
+            {
+                "role": "user",
+                "content": [
+                    {"type": "text", "text": prompt},
+                    {
+                        "type": "image_url",
+                        "image_url": {
+                            "url": f"data:image/png;base64,{base64_image}"
+                        }
+                    }
+                ]
+            }
+        ],
+        max_tokens=1000
+    )
+
+    return response.choices[0].message.content
+
+# Define the args schema for reflect_on_image
+class ReflectOnImageArgs(BaseModel):
+    image_path: str = Field(description="The path to the image to reflect on.")
+
+# Define the reflect_on_image tool
+reflect_tool = StructuredTool.from_function(
+    func=reflect_on_image,
+    name="reflect_on_image",
+    description="A tool to reflect on an image and provide feedback for improvements.",
+    args_schema=ReflectOnImageArgs
+)
